requirements.txt
discord.py==2.3.2
Pillow==10.0.0
colorama==0.4.6
termcolor==2.3.0
pyfiglet==0.8.post1